Item feed for mobile, fix up
