+++ src/app/item-feed/Highlights.m.scss
@@ -1,0 +1,34 @@
+@import '../variables.scss';
+
+.type {
+  > * {
+    display: inline;
+  }
+}
+
+.perks {
+  display: flex;
+  flex-flow: column;
+  margin: 4px 0;
+
+  > div > div {
+    display: flex;
+    flex-flow: row;
+    align-items: center;
+  }
+
+  img {
+    height: 24px;
+    width: 24px;
+  }
+}
+
+.stats {
+  margin: 4px 0 0 0;
+  :global(.stat-row) {
+    margin-bottom: 8px;
+  }
+  :global(.stat) {
+    line-height: 8px;
+  }
+}
