+++ src/app/item-feed/Highlights.m.scss.d.ts
@@ -1,0 +1,9 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'perks': string;
+  'stats': string;
+  'type': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
