+++ src/app/item-feed/ItemFeed.m.scss.d.ts
@@ -2,17 +2,9 @@
 // Please do not change this file!
 interface CssExports {
   'classIcon': string;
-  'expanded': string;
   'info': string;
   'item': string;
-  'perks': string;
-  'sideTray': string;
-  'stats': string;
-  'tagButton': string;
   'title': string;
-  'trayButton': string;
-  'trayContainer': string;
-  'type': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
