+++ src/app/item-feed/ItemFeedPage.m.scss
@@ -1,0 +1,4 @@
+.page {
+  composes: dim-page from global;
+  padding: 10px;
+}
