+++ src/app/item-feed/ItemFeedPage.m.scss.d.ts
@@ -1,0 +1,7 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'page': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
