+++ src/app/item-feed/ItemFeedSidebar.m.scss.d.ts
@@ -1,0 +1,10 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'expanded': string;
+  'sideTray': string;
+  'trayButton': string;
+  'trayContainer': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
