+++ src/app/item-feed/TagButtons.m.scss.d.ts
@@ -1,0 +1,7 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'tagButton': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
